@@ -4,7 +4,7 @@ class AutoDeployAgentException(Exception):
     def __init__(self, msg=None):
         if msg is None:
             msg = "Unexpected error was triggered"
-        super(AutoDeployAgentException, self).__init__(msg)
+        super().__init__(msg)
 
 
 class NoErrorException(AutoDeployAgentException):
@@ -12,7 +12,7 @@ class NoErrorException(AutoDeployAgentException):
 
     def __init__(self):
         msg = "no error, no problem"
-        super(NoErrorException, self).__init__(msg)
+        super().__init__(msg)
 
 
 class CmdLineError(AutoDeployAgentException):
@@ -20,8 +20,9 @@ class CmdLineError(AutoDeployAgentException):
 
     def __init__(self, cmd, error_msg):
         msg = "Error while executing {cmd}:\n{error_msg}".format(
-            cmd=cmd, error_msg=error_msg)
-        super(CmdLineError, self).__init__(msg)
+            cmd=cmd, error_msg=error_msg
+        )
+        super().__init__(msg)
         self.cmd = cmd
         self.error_msg = error_msg
 
@@ -30,11 +31,11 @@ class ConfigurationError(AutoDeployAgentException):
     """Wrong configuration error"""
 
     def __init__(self, msg):
-        super(ConfigurationError, self).__init__(msg)
+        super().__init__(msg)
 
 
 class DependencyNotReadyError(AutoDeployAgentException):
     """Dependency not ready error"""
 
     def __init__(self, msg):
-        super(DependencyNotReadyError, self).__init__(msg)
+        super().__init__(msg)
