@@ -9,7 +9,8 @@ from servicelib.rest_responses import wrap_as_envelope
 from servicelib.rest_utils import body_to_dict, extract_and_validate
 
 from . import __version__
-from .auto_deploy_task import TASK_STATE
+from .app_state import State
+from .auto_deploy_task import TASK_NAME
 
 log = logging.getLogger(__name__)
 
@@ -21,12 +22,12 @@ async def check_health(request: web.Request):
     assert not params
     assert not query
     assert not body
-
+    app_state = app.get("state", {TASK_NAME: State.FAILED})
     data = {
-        'name': __name__.split('.')[0],
-        'version': __version__,
-        'status': f"SERVICE_{app[TASK_STATE].name}" if TASK_STATE in app else "SERVICE_FAILED",
-        'api_version': __version__
+        "name": __name__.split(".")[0],
+        "version": __version__,
+        "status": f"SERVICE_{app_state[TASK_NAME].name}",
+        "api_version": __version__,
     }
 
     return data
@@ -39,14 +40,14 @@ async def check_action(request: web.Request):
     assert query, "query %s" % query
     assert body, "body %s" % body
 
-    if params['action'] == 'fail':
+    if params["action"] == "fail":
         raise ValueError("some randome failure")
 
     # echo's input
     data = {
-        "path_value": params.get('action'),
-        "query_value": query.get('data'),
-        "body_value": body_to_dict(body)
+        "path_value": params.get("action"),
+        "query_value": query.get("data"),
+        "body_value": body_to_dict(body),
     }
 
     return wrap_as_envelope(data=data)
